feat(plugin): expose active listener URL
