+++ packages/plugin/src/auth.ts
@@ -1,0 +1,133 @@
+import type { Provider } from "@opencode-ai/sdk"
+import type { Auth } from "@opencode-ai/sdk/v2"
+
+type Rule = {
+  key: string
+  op: "eq" | "neq"
+  value: string
+}
+
+export type AuthHook = {
+  provider: string
+  loader?: (auth: () => Promise<Auth>, provider: Provider) => Promise<Record<string, any>>
+  methods: (
+    | {
+        type: "oauth"
+        label: string
+        prompts?: Array<
+          | {
+              type: "text"
+              key: string
+              message: string
+              placeholder?: string
+              validate?: (value: string) => string | undefined
+              /** @deprecated Use `when` instead */
+              condition?: (inputs: Record<string, string>) => boolean
+              when?: Rule
+            }
+          | {
+              type: "select"
+              key: string
+              message: string
+              options: Array<{
+                label: string
+                value: string
+                hint?: string
+              }>
+              /** @deprecated Use `when` instead */
+              condition?: (inputs: Record<string, string>) => boolean
+              when?: Rule
+            }
+        >
+        authorize(inputs?: Record<string, string>): Promise<AuthOAuthResult>
+      }
+    | {
+        type: "api"
+        label: string
+        prompts?: Array<
+          | {
+              type: "text"
+              key: string
+              message: string
+              placeholder?: string
+              validate?: (value: string) => string | undefined
+              /** @deprecated Use `when` instead */
+              condition?: (inputs: Record<string, string>) => boolean
+              when?: Rule
+            }
+          | {
+              type: "select"
+              key: string
+              message: string
+              options: Array<{
+                label: string
+                value: string
+                hint?: string
+              }>
+              /** @deprecated Use `when` instead */
+              condition?: (inputs: Record<string, string>) => boolean
+              when?: Rule
+            }
+        >
+        authorize?(inputs?: Record<string, string>): Promise<
+          | {
+              type: "success"
+              key: string
+              provider?: string
+              metadata?: Record<string, string>
+            }
+          | {
+              type: "failed"
+            }
+        >
+      }
+  )[]
+}
+
+export type AuthOAuthResult = { url: string; instructions: string } & (
+  | {
+      method: "auto"
+      callback(): Promise<
+        | ({
+            type: "success"
+            provider?: string
+          } & (
+            | {
+                refresh: string
+                access: string
+                expires: number
+                accountId?: string
+                enterpriseUrl?: string
+              }
+            | { key: string; metadata?: Record<string, string> }
+          ))
+        | {
+            type: "failed"
+          }
+      >
+    }
+  | {
+      method: "code"
+      callback(code: string): Promise<
+        | ({
+            type: "success"
+            provider?: string
+          } & (
+            | {
+                refresh: string
+                access: string
+                expires: number
+                accountId?: string
+                enterpriseUrl?: string
+              }
+            | { key: string; metadata?: Record<string, string> }
+          ))
+        | {
+            type: "failed"
+          }
+      >
+    }
+)
+
+/** @deprecated Use AuthOAuthResult instead. */
+export type AuthOuathResult = AuthOAuthResult
