+++ packages/plugin/src/config.ts
@@ -1,0 +1,7 @@
+import type { Config as SDKConfig } from "@opencode-ai/sdk"
+
+export type PluginOptions = Record<string, unknown>
+
+export type Config = Omit<SDKConfig, "plugin"> & {
+  plugin?: Array<string | [string, PluginOptions]>
+}
